Refactor dependency URL handling in telemetry
* Avoid back and forth conversions between URL, String, URI, just
  use URL directly and do early checks to enforce it's either file:
  or jar:file: by the time it is enqueued.
* Remove code redundancies, and move dependency resolving logic to
  DependencyResolver.
